+++ Ullmannite/src/UiElement/Controls/UiRectGradient.cpp
@@ -89,7 +89,7 @@ void UiRectGradient::CreateResources()
 
     m_layout->Bind();
 
-    m_vertexBuffer = VertexBuffer::Create(sizeof(VertexDesc) * vertices.size(), reinterpret_cast<float*>(vertices.data()), GraphicsBufferType::STATIC_DRAW);
+    m_vertexBuffer = VertexBuffer::Create(sizeof(VertexDesc) * static_cast<int>(vertices.size()), reinterpret_cast<float*>(vertices.data()), GraphicsBufferType::STATIC_DRAW);
     m_indexBuffer = IndexBuffer::Create(sizeof(indices), indices, GraphicsBufferType::STATIC_DRAW);
 
     m_layout->Build();
